pyproject.toml
[tool.poetry]
name = "pest-cli"
version = "0.1.0"
description = "🐀 ⇝ cli tool for pest applications"
authors = ["Lucas Colombo <lucas@lucaslabs.tech>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
inquirerpy = "^0.3.4"
pygments = "^2.17.2"
pyyaml = "^6.0.1"
click = "^8.1.7"
rich = "^13.7.0"
dataclass-wizard = "^0.22.2"
jinja2 = "^3.1.2"
tomli = "^2.0.1"
typing-extensions = "^4.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
ruff = "^0.1.6"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
nox-poetry = "^1.0.3"
black = "^23.12.1"


[tool.poetry.scripts]
pest = "pest_cli.app:cli"

[tool.black]
line-length = 100
skip-string-normalization = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

